+++ src/Sfa.Das.Sas.Indexer.UnitTests/MetaDataCreationTool/Services/AssessmentOrgsExcelPackageServiceTests.cs
@@ -6,13 +6,13 @@ namespace Sfa.Das.Sas.Indexer.UnitTests.MetaDataCreationTool.Services
     [TestFixture]
     public class AssessmentOrgsExcelPackageServiceTests
     {
-        [TestCase("11112222", 11112222)]
-        [TestCase("011122222", 11122222)]
+        [TestCase("11112222", 11112222L)]
+        [TestCase("011122222", 11122222L)]
         [TestCase("1111222", null)]
         [TestCase("111122222", null)]
         [TestCase("", null)]
         [TestCase("-", null)]
-        public void ShouldProcessUkprnStringToReturnOnlyValidUkprns(string ukprnInput, int? expected)
+        public void ShouldProcessUkprnStringToReturnOnlyValidUkprns(string ukprnInput, long? expected)
         {
             var sut = new AssessmentOrgsExcelPackageService();
             var actual = sut.CheckForValidUkprn(ukprnInput);
